pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcportal"
version = "0.1.0"
description = "A Flask-based Minecraft server control panel with approvals and audit logging."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Flask>=3.1,<4.0",
  "Flask-SQLAlchemy>=3.1,<4.0",
  "Flask-Login>=0.6,<1.0",
  "Flask-WTF>=1.2,<2.0",
  "Flask-Limiter>=3.10,<4.0",
  "Flask-Talisman>=1.1,<2.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "bandit[toml]>=1.7,<2.0",
  "pip-audit>=2.8,<3.0",
  "pytest>=8.3,<9.0",
  "pytest-cov>=6.0,<7.0",
  "ruff>=0.11,<0.12",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
